% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwc_simple.R
\name{darwin_simple}
\alias{darwin_simple}
\title{Construct simple darwin core}
\usage{
darwin_simple(..., .list = list())
}
\arguments{
\item{..., .list}{Either named params or a named list, not both}
}
\value{
an \code{xml_document}
}
\description{
Construct simple darwin core
}
\examples{
library(xml2)
# just dublin core
out <- darwin_simple(type = "PhysicalObject", modified = "2009-02-12T14:22:00Z", language = "en")
out
xml_structure(out)
write_xml(out, (f <- tempfile(fileext=".xml")))

# add some darwin core
out <- darwin_simple(basisOfRecord = "PreservedSpecimen", taxonID = "12345", scientificName = "Poa annua")
out
xml_structure(out)
write_xml(out, (f <- tempfile(fileext=".xml")))

# together
out <- darwin_simple(type = "PhysicalObject", 
  modified = "2009-02-12", language = "en", 
  basisOfRecord = "PreservedSpecimen", taxonID = "12345", 
  scientificName = "Poa annua")
out
xml_structure(out)
write_xml(out, (f <- tempfile(fileext=".xml")))
}
