% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{tdc_init}
\alias{tdc_init}
\title{Initialize}
\usage{
tdc_init(x)
}
\arguments{
\item{x}{a thing}
}
\description{
Initialize
}
\examples{
\dontrun{
library(taxa)
x <- taxonomy(c("mammalia", "felidae", "panthera", "tigris"),
              c("mammalia", "felidae", "panthera", "leo"),
              c("mammalia", "felidae", "felis", "catus"))
tdc_init(x)
}
}
