% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdc_tax_strings.R
\name{tdc_tax_strings}
\alias{tdc_tax_strings}
\title{get tax strings}
\usage{
tdc_tax_strings(x)
}
\arguments{
\item{x}{input, a Taxonomy object for now only}
}
\description{
get tax strings
}
\examples{
library(taxa)
ex_taxonomy
tdc_tax_strings(x = ex_taxonomy)
}
