% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwin_simple.R
\name{darwin_simple}
\alias{darwin_simple}
\title{Construct a simple darwin core list}
\usage{
darwin_simple(..., .list = list())
}
\arguments{
\item{..., .list}{Either named params or a named list, not both}
}
\value{
a list of class \code{darwin_simple}
}
\description{
Construct a simple darwin core list
}
\examples{
# just dublin core
out <- darwin_simple(type = "PhysicalObject", 
  modified = "2009-02-12", language = "en")
out

# add some darwin core
out <- darwin_simple(basisOfRecord = "PreservedSpecimen", 
  taxonID = "12345", scientificName = "Poa annua")
out

# together
out <- darwin_simple(type = "PhysicalObject", 
  modified = "2009-02-12", language = "en", 
  basisOfRecord = "PreservedSpecimen", taxonID = "12345", 
  scientificName = "Poa annua")
out
}
